Fixed message types and switched over to activation event

+++ src/app/util/Message.cpp
@@ -2,12 +2,12 @@
 
 namespace autom8 { namespace app { namespace message {
 
-    int64_t CreateType() {
-        static int64_t nextMessageType = 2048;
+    int CreateType() {
+        static int nextMessageType = 2048;
         return ++nextMessageType;
     }
 
-    const int64_t BROADCAST_SWITCH_TO_CLIENT_LAYOUT = CreateType();
-    const int64_t BROADCAST_SWITCH_TO_SETTINGS_LAYOUT = CreateType();
+    const int BROADCAST_SWITCH_TO_CLIENT_LAYOUT = CreateType();
+    const int BROADCAST_SWITCH_TO_SETTINGS_LAYOUT = CreateType();
 
 }}}
