+++ src/app/util/Message.h
@@ -5,9 +5,9 @@
 
 namespace autom8 { namespace app { namespace message {
 
-    int64_t CreateType();
+    int CreateType();
 
-    extern const int64_t BROADCAST_SWITCH_TO_CLIENT_LAYOUT;
-    extern const int64_t BROADCAST_SWITCH_TO_SETTINGS_LAYOUT;
+    extern const int BROADCAST_SWITCH_TO_CLIENT_LAYOUT;
+    extern const int BROADCAST_SWITCH_TO_SETTINGS_LAYOUT;
 
 } } }
